GEOS data added
